requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
pytz==2024.1